+++ tests/unit/apis/test_queue.py
@@ -77,7 +77,7 @@ def test_it_adds_to_queue(table):
         },
         SimpleNamespace(),
     )
-
+    print(response["body"])
     assert 201 == response["statusCode"]
     assert {
         "DeletionQueueItemId": ANY,