+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## v0.11 (unreleased)
+## v0.11
 
 - [#200](https://github.com/awslabs/amazon-s3-find-and-forget/pull/200): Add API
   Endpoint for adding deletion queue items in batch - deprecates PATCH /v1/queue